Refactored notifications app into different JS files.
