+++ src/front/apps/fl-notifications/comments/index.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { ContentQuery } from 'components'
-import { NotificationsTabEmptyMessage } from '../empty-message'
 import { commentsQuery } from './queries'
 
 export const Comments = () => {
@@ -9,7 +8,6 @@ export const Comments = () => {
 			type={ 'comments' }
 			query={ commentsQuery() }
 			pagination={ true }
-			emptyMessage={ <NotificationsTabEmptyMessage /> }
 		/>
 	)
 }
