+++ src/front/apps/fl-notifications/updates/empty-message.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { EmptyMessage } from 'components'
 
-export const NotificationsTabEmptyMessage = () => {
+export const UpdatesEmptyMessage = () => {
 	return <EmptyMessage>You're all up to date!</EmptyMessage>
 }
