+++ src/front/apps/fl-notifications/updates/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ContentQuery } from 'components'
 import { UpdatesListItem } from './list-item'
-import { NotificationsTabEmptyMessage } from '../empty-message'
+import { UpdatesEmptyMessage } from './empty-message'
 import { updatesQuery } from './queries'
 
 export const Updates = () => {
@@ -11,7 +11,7 @@ export const Updates = () => {
 			query={ updatesQuery() }
 			pagination={ false }
 			item={ <UpdatesListItem /> }
-			emptyMessage={ <NotificationsTabEmptyMessage /> }
+			emptyMessage={ <UpdatesEmptyMessage /> }
 		/>
 	)
 }
